Fixes bugs (#7409)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
fix: Fixes a variety of bugs and incorrect span usage.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

+++ code/game/rotational_objects/item_rotation_contraption.dm
@@ -68,20 +68,22 @@
 	placed_type = parent_type
 
 /obj/item/rotation_contraption/interact_with_atom(atom/interacting_with, mob/living/user, list/modifiers)
-	if(istype(interacting_with, /obj/item/rotation_contraption))
-		if(!can_stack)
-			return
+	if(isitem(interacting_with))
+		if(istype(interacting_with, /obj/item/rotation_contraption))
+			if(!can_stack)
+				return
 
-		var/obj/item/rotation_contraption/rotator = interacting_with
-		if(placed_type != rotator.placed_type)
-			return
+			var/obj/item/rotation_contraption/rotator = interacting_with
+			if(placed_type != rotator.placed_type)
+				return
 
-		in_stack += rotator.in_stack
-		balloon_alert(user, "stacked!")
-		update_appearance(UPDATE_NAME)
-		qdel(rotator)
+			in_stack += rotator.in_stack
+			balloon_alert(user, "stacked!")
+			update_appearance(UPDATE_NAME)
+			qdel(rotator)
 
-		return ITEM_INTERACT_SUCCESS
+			return ITEM_INTERACT_SUCCESS
+		return ..()
 
 	var/turf/T = get_turf(interacting_with)
 
@@ -103,7 +105,8 @@
 			return ITEM_INTERACT_BLOCKING
 
 	visible_message("[user] starts placing down [src].", "You start to place [src].")
-	if(!do_after(user, 1.2 SECONDS - GET_MOB_SKILL_VALUE_OLD(user, /datum/attribute/skill/craft/engineering), T))
+	var/delay = max(1 SECONDS - GET_MOB_SKILL_VALUE_OLD(user, /datum/attribute/skill/craft/engineering), 0)
+	if(!do_after(user, delay, T))
 		return ITEM_INTERACT_BLOCKING
 
 	var/obj/structure/structure = new placed_type(T)
